move sessionEpireListener to own class. refs #1698
also comment out unused parameters in `core.xml`
